Extract getActiveSubscriptions
